Create settings-menu-mixin for subview state tracking
Ref #4642
- Move duplicate code from PSM and SettingsTags to a mixin
